Create version config on push.py (#70)
push.py now prompts for creating the version configuration for this game
version if CREATE_VERSION_CONFIG is true. It adds the config to the top
to ensure order is retained and has a confirmation step.

Git tagging the commit has been removed as the data repo's summarizer
workflow now uses just two versions as input and works off the archive.
As such, the workflow is now only triggered if pushing to archive.